python-client/requirements.txt
# MCP Client dependencies
mcp>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
